requirements.txt
Flask==3.1.0
pandas==2.2.2
python-dotenv==1.0.1
